Added Android ndk-build support
Starting with Android Studio 2.2 ndk-build supports CMake as an external
build system to build native libraries.

This commit adds support for building openssl-cmake with ndk-build,
respecting the CMake build environment. Due to problems with OpenSSL's
platform auto detection when using ndk-build (mostly issues with using
clang as the compiler) we cannot call convenient config script but must
use the verbose and low-level Configure.  But there is still one
problem: OpenSSL supplies "-mandroid" as compiler parameter. Clang does
not understand this parameter and treats it as an error as of version
3.4. Silencing the error with
-Wno-error=unused-command-line-argument-hard-error-in-future is not
possible anymore. To solve this problem we supply a patch to OpenSSL
sources until the issue is fixed upstream
(https://github.com/openssl/openssl/issues/1531).